Rename the directory holding the node.
We ran into an interesting issue/potential Python bug.  If you
have multiple load_entry scripts that have the same *intermediate*
directory name, then the load_entry stuff tries to 'import intermediate'
before anything else.  The problem there is that it always gets
the "first" one (I'm not sure of the resolution order).  Thus,
to work around this for right now, we rename nodes to jpeg_to_raw,
which gets rid of this problem.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>